🔨 Refactor(#338): 신청서 제출 페이지 리팩터링
