+++ src/app/(steady)/(application)/application/submit/[steady_id]/page.tsx
@@ -4,27 +4,25 @@ import type { ChangeEvent } from "react";
 import { useEffect, useState } from "react";
 import { usePathname, useRouter } from "next/navigation";
 import { Question, Title } from "@/components/application";
-import { useToast } from "@/components/ui/use-toast";
 import { TextArea } from "@radix-ui/themes";
 import { useSuspenseQuery } from "@tanstack/react-query";
-import submitApplication from "@/services/application/submitApplication";
 import getSteadyDetails from "@/services/steady/getSteadyDetails";
 import getSteadyQuestions from "@/services/steady/getSteadyQuestions";
-import Button, { buttonSize } from "@/components/_common/Button";
+import SubmitButtons from "@/components/containers/applicationSubmit/SubmitButtons";
 import {
   getSteadyDetailsKey,
   getSteadyQuestionsKey,
 } from "@/constants/queryKeys";
 
-interface pageParams {
-  id: string;
-}
-
-const ApplicationSubmitPage = ({ params }: { params: pageParams }) => {
-  const router = useRouter();
+const ApplicationSubmitPage = ({
+  params,
+}: {
+  params: { steady_id: string };
+}) => {
   const pathname = usePathname();
-  const steadyId = params.id;
+  const steadyId = params.steady_id;
   const pageType = pathname.split("/")[2];
+  const router = useRouter();
   const { data: steadyQuestionsData } = useSuspenseQuery({
     queryKey: getSteadyQuestionsKey(steadyId),
     queryFn: () => getSteadyQuestions(steadyId),
@@ -36,14 +34,12 @@ const ApplicationSubmitPage = ({ params }: { params: pageParams }) => {
       answer: "",
     })),
   );
-  const { toast } = useToast();
 
-  const { data: steadyDetailsData, refetch: steadyDetailsRefetch } =
-    useSuspenseQuery({
-      queryKey: getSteadyDetailsKey(steadyId),
-      queryFn: () => getSteadyDetails(steadyId),
-      staleTime: 10000,
-    });
+  const { data: steadyDetailsData } = useSuspenseQuery({
+    queryKey: getSteadyDetailsKey(steadyId),
+    queryFn: () => getSteadyDetails(steadyId),
+    staleTime: 10000,
+  });
 
   useEffect(() => {
     if (steadyDetailsData.applicationId !== null) {
@@ -61,32 +57,6 @@ const ApplicationSubmitPage = ({ params }: { params: pageParams }) => {
         id === index ? { ...question, answer: newAnswer } : question,
       ),
     );
-  };
-
-  const onSubmitAnswers = async () => {
-    const checkInvalidAnswers = applicationData.some(
-      (item) => !item.answer.length,
-    );
-    if (checkInvalidAnswers) {
-      toast({
-        description: "입력되지 않은 항목이 있습니다.",
-        variant: "red",
-      });
-    } else {
-      try {
-        await submitApplication(steadyId, applicationData);
-        toast({
-          description: "스테디 참여 신청서 제출 성공!",
-          variant: "green",
-        });
-        steadyDetailsRefetch();
-      } catch (error) {
-        toast({
-          description: "스테디 참여 신청서 제출 실패!",
-          variant: "red",
-        });
-      }
-    }
   };
 
   return (
@@ -117,18 +87,10 @@ const ApplicationSubmitPage = ({ params }: { params: pageParams }) => {
         ))}
       </Title>
       <div className="flex items-center justify-end gap-20">
-        <Button
-          className={`${buttonSize.sm} bg-st-white`}
-          onClick={() => router.back()}
-        >
-          취소
-        </Button>
-        <Button
-          onClick={onSubmitAnswers}
-          className={`${buttonSize.sm} bg-st-primary text-st-white`}
-        >
-          제출 완료
-        </Button>
+        <SubmitButtons
+          applicationData={applicationData}
+          steadyId={steadyId}
+        />
       </div>
     </>
   );
